Move unix-specific syscalls and file descriptors out of os-agnostic files
